examples/analytics-api/python/requirements.txt
# TESAIoT Analytics API Client Dependencies

# HTTP client with async support
httpx>=0.25.0

# Data validation
pydantic>=2.0.0

# Date/time handling
python-dateutil>=2.8.0

# Environment variables
python-dotenv>=1.0.0

# Optional: for dashboard example
rich>=13.0.0
